Update API docs to new layout
